+++ lib/main.dart
@@ -1,968 +1,24 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'dart:async';
 import 'firebase_options.dart';
+import 'screens/splash_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
-);
+  );
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: const SplashScreen(),
-    );
-  }
-}
-
-class SplashScreen extends StatefulWidget {
-  const SplashScreen({super.key});
-
-  @override
-  State<SplashScreen> createState() => _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen> {
-  @override
-  void initState() {
-    super.initState();
-    Future.delayed(const Duration(milliseconds: 1500), () {
-      if (!mounted) return;
-      Navigator.of(context).pushReplacement(
-        MaterialPageRoute(builder: (_) => const OnboardingScreen()),
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      backgroundColor: Color(0xFF201E1A),
-      body: Center(
-        child: Image(
-          image: AssetImage('assets/logo.png'),
-          fit: BoxFit.contain,
-        ),
-      ),
-    );
-  }
-}
-
-class OnboardingScreen extends StatelessWidget {
-  const OnboardingScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 12),
-              Expanded(
-                child: Align(
-                  alignment: Alignment.topLeft,
-                  child: Image.asset(
-                    'assets/onboarding_1.png',
-                    fit: BoxFit.contain,
-                  ),
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Explore Health',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: Colors.white,
-                  shadows: const [
-                    Shadow(color: Colors.black54, blurRadius: 6),
-                  ],
-                ),
-              ),
-              Text(
-                'Together',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: const Color(0xFFCDEDC6),
-                  shadows: const [
-                    Shadow(color: Colors.black54, blurRadius: 6),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 12),
-              Text(
-                'Join a community crowdsourcing\nreal experiments for better health\ninsights.',
-                style: TextStyle(
-                  color: Colors.white.withOpacity(0.75),
-                  fontSize: 14,
-                  height: 1.5,
-                ),
-              ),
-              const SizedBox(height: 28),
-              SizedBox(
-                width: double.infinity,
-                child: OutlinedButton.icon(
-                  onPressed: () {
-                    Navigator.of(context).push(
-                      _slideRoute(const OnboardingScreenTwo()),
-                    );
-                  },
-                  icon: const Icon(Icons.arrow_right_alt_rounded, size: 24),
-                  label: const Text("Let's Get Started"),
-                  style: OutlinedButton.styleFrom(
-                    foregroundColor: Colors.white,
-                    side: BorderSide(color: Colors.white.withOpacity(0.25)),
-                    padding: const EdgeInsets.symmetric(vertical: 18, horizontal: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(28),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 20),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Creates a right-to-left slide page transition.
-PageRouteBuilder _slideRoute(Widget page) {
-  return PageRouteBuilder(
-    transitionDuration: const Duration(milliseconds: 350),
-    pageBuilder: (context, animation, secondaryAnimation) => page,
-    transitionsBuilder: (context, animation, secondaryAnimation, child) {
-      final tween = Tween<Offset>(begin: const Offset(1.0, 0.0), end: Offset.zero)
-          .chain(CurveTween(curve: Curves.easeOutCubic));
-      return SlideTransition(position: animation.drive(tween), child: child);
-    },
-  );
-}
-
-class OnboardingScreenTwo extends StatelessWidget {
-  const OnboardingScreenTwo({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 12),
-              Expanded(
-                child: Align(
-                  alignment: Alignment.topLeft,
-                  child: Image.asset(
-                    'assets/onboarding_2.png',
-                    fit: BoxFit.contain,
-                  ),
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Log Your',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: Colors.white,
-                  shadows: const [Shadow(color: Colors.black54, blurRadius: 6)],
-                ),
-              ),
-              Text(
-                'Experiments',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: const Color(0xFFCDEDC6),
-                  shadows: const [Shadow(color: Colors.black54, blurRadius: 6)],
-                ),
-              ),
-              const SizedBox(height: 12),
-              Text(
-                'Track habits, routines, or tests and\ncontribute to collective discoveries.',
-                style: TextStyle(
-                  color: Colors.white.withOpacity(0.75),
-                  fontSize: 14,
-                  height: 1.5,
-                ),
-              ),
-              const SizedBox(height: 28),
-              SizedBox(
-                width: double.infinity,
-                child: OutlinedButton.icon(
-                  onPressed: () {
-                    Navigator.of(context).push(
-                      _slideRoute(const OnboardingScreenThree()),
-                    );
-                  },
-                  icon: const Icon(Icons.arrow_right_alt_rounded, size: 24),
-                  label: const Text('Next'),
-                  style: OutlinedButton.styleFrom(
-                    foregroundColor: Colors.white,
-                    side: BorderSide(color: Colors.white.withOpacity(0.25)),
-                    padding: const EdgeInsets.symmetric(vertical: 18, horizontal: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(28),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 20),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class OnboardingScreenThree extends StatelessWidget {
-  const OnboardingScreenThree({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 12),
-              Expanded(
-                child: Align(
-                  alignment: Alignment.topLeft,
-                  child: Image.asset(
-                    'assets/onboarding_3.png',
-                    fit: BoxFit.contain,
-                  ),
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Discover',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: Colors.white,
-                  shadows: const [Shadow(color: Colors.black54, blurRadius: 6)],
-                ),
-              ),
-              Text(
-                'Insights',
-                style: TextStyle(
-                  fontSize: 36,
-                  fontWeight: FontWeight.w800,
-                  color: const Color(0xFFCDEDC6),
-                  shadows: const [Shadow(color: Colors.black54, blurRadius: 6)],
-                ),
-              ),
-              const SizedBox(height: 12),
-              Text(
-                'See trends and findings tailored to you\nfrom community experiments.',
-                style: TextStyle(
-                  color: Colors.white.withOpacity(0.75),
-                  fontSize: 14,
-                  height: 1.5,
-                ),
-              ),
-              const SizedBox(height: 28),
-              SizedBox(
-                width: double.infinity,
-                child: OutlinedButton.icon(
-                  onPressed: () {
-                    Navigator.of(context).push(
-                      _slideRoute(const SignUpScreen()),
-                    );
-                  },
-                  icon: const Icon(Icons.arrow_right_alt_rounded, size: 24),
-                  label: const Text('Get Started'),
-                  style: OutlinedButton.styleFrom(
-                    foregroundColor: Colors.white,
-                    side: BorderSide(color: Colors.white.withOpacity(0.25)),
-                    padding: const EdgeInsets.symmetric(vertical: 18, horizontal: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(28),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 20),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ContentPreferenceScreen extends StatefulWidget {
-  const ContentPreferenceScreen({super.key});
-
-  @override
-  State<ContentPreferenceScreen> createState() => _ContentPreferenceScreenState();
-}
-
-class _ContentPreferenceScreenState extends State<ContentPreferenceScreen> {
-  final List<_PreferenceItem> _items = const <_PreferenceItem>[
-    _PreferenceItem(label: 'Gym & Strength', assetPath: 'assets/prefs/gym.png'),
-    _PreferenceItem(label: 'Nutrition & Food', assetPath: 'assets/prefs/nutrition.png'),
-    _PreferenceItem(label: 'Sleep & Recovery', assetPath: 'assets/prefs/sleep.png'),
-    _PreferenceItem(label: 'Mental Wellness', assetPath: 'assets/prefs/mental.png'),
-    _PreferenceItem(label: 'Daily Exercise', assetPath: 'assets/prefs/exercise.png'),
-    _PreferenceItem(label: 'Energy & Focus', assetPath: 'assets/prefs/energy.png'),
-    _PreferenceItem(label: 'Heart & Health', assetPath: 'assets/prefs/heart.png'),
-    _PreferenceItem(label: 'Hydration', assetPath: 'assets/prefs/hydration.png'),
-    _PreferenceItem(label: 'Supplements', assetPath: 'assets/prefs/supplements.png'),
-  ];
-
-  final Set<int> _selectedIndexes = <int>{};
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-        child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-              const SizedBox(height: 12),
-              Text(
-                'Choose your',
-                style: const TextStyle(
-                  fontSize: 32,
-                  fontWeight: FontWeight.w800,
-                  color: Colors.white,
-                ),
-              ),
-              Text(
-                'Interests',
-                style: const TextStyle(
-                  fontSize: 32,
-                  fontWeight: FontWeight.w800,
-                  color: Colors.white,
-                ).copyWith(shadows: const [Shadow(color: Colors.black54, blurRadius: 6)]),
-              ),
-              const SizedBox(height: 8),
-            Text(
-                "Pick what you'd like to explore in\nHealthLab. We'll personalize your\nexperiments",
-                style: TextStyle(
-                  color: Colors.white.withOpacity(0.75),
-                  fontSize: 14,
-                  height: 1.5,
-                ),
-              ),
-              const SizedBox(height: 16),
-              Expanded(
-                child: GridView.builder(
-                  padding: EdgeInsets.zero,
-                  gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                    crossAxisCount: 3,
-                    mainAxisSpacing: 16,
-                    crossAxisSpacing: 16,
-                    childAspectRatio: 0.9,
-                  ),
-                  itemCount: _items.length,
-                  itemBuilder: (context, index) {
-                    final item = _items[index];
-                    final isSelected = _selectedIndexes.contains(index);
-                    return _PreferenceTile(
-                      item: item,
-                      selected: isSelected,
-                      onTap: () {
-    setState(() {
-                          if (isSelected) {
-                            _selectedIndexes.remove(index);
-                          } else {
-                            _selectedIndexes.add(index);
-                          }
-                        });
-                      },
-                    );
-                  },
-                ),
-              ),
-              const SizedBox(height: 8),
-              SizedBox(
-                width: double.infinity,
-                child: OutlinedButton.icon(
-                  onPressed: () {
-                    // TODO: Wire to next screen or home when backend ready
-                  },
-                  icon: const Icon(Icons.arrow_right_alt_rounded, size: 24),
-                  label: const Text('Next'),
-                  style: OutlinedButton.styleFrom(
-                    foregroundColor: Colors.white,
-                    side: BorderSide(color: Colors.white.withOpacity(0.25)),
-                    padding: const EdgeInsets.symmetric(vertical: 18, horizontal: 20),
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(28),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 20),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class SignUpScreen extends StatefulWidget {
-  const SignUpScreen({super.key});
-
-  @override
-  State<SignUpScreen> createState() => _SignUpScreenState();
-}
-
-class _SignUpScreenState extends State<SignUpScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _nameController = TextEditingController();
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  final TextEditingController _confirmPasswordController = TextEditingController();
-
-  bool _submitting = false;
-
-  Future<void> _registerWithEmail() async {
-    if (!_formKey.currentState!.validate()) return;
-    setState(() => _submitting = true);
-    try {
-      final email = _emailController.text.trim();
-      final password = _passwordController.text;
-      await FirebaseAuth.instance.createUserWithEmailAndPassword(email: email, password: password);
-      final user = FirebaseAuth.instance.currentUser;
-      await user?.sendEmailVerification();
-      if (!mounted) return;
-      Navigator.of(context).pushReplacement(
-        _slideRoute(EmailVerificationScreen(email: email)),
-      );
-    } on FirebaseAuthException catch (e) {
-      if (!mounted) return;
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Sign up failed')));
-    } finally {
-      if (mounted) setState(() => _submitting = false);
-    }
-  }
-
-  Future<void> _continueWithGoogle() async {
-    setState(() => _submitting = true);
-    try {
-      final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
-      if (googleUser == null) {
-        setState(() => _submitting = false);
-        return; // canceled
-      }
-      final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
-      final credential = GoogleAuthProvider.credential(
-        accessToken: googleAuth.accessToken,
-        idToken: googleAuth.idToken,
-      );
-      await FirebaseAuth.instance.signInWithCredential(credential);
-      if (!mounted) return;
-      Navigator.of(context).pushReplacement(_slideRoute(const UsernameScreen()));
-    } on FirebaseAuthException catch (e) {
-      if (!mounted) return;
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Google sign-in failed')));
-    } finally {
-      if (mounted) setState(() => _submitting = false);
-    }
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _emailController.dispose();
-    _passwordController.dispose();
-    _confirmPasswordController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: SingleChildScrollView(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 16),
-              const Text(
-                'Welcome!',
-                style: TextStyle(fontSize: 32, fontWeight: FontWeight.w800, color: Colors.white),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                "Let's create you an account",
-                style: TextStyle(color: Colors.white.withOpacity(0.7)),
-              ),
-              const SizedBox(height: 24),
-              Form(
-                key: _formKey,
-                child: Column(
-                  children: <Widget>[
-                    _Field(label: 'Full name', controller: _nameController, keyboardType: TextInputType.name),
-                    const SizedBox(height: 14),
-                    _Field(label: 'Email', controller: _emailController, keyboardType: TextInputType.emailAddress,
-                      validator: (v){
-                        if (v == null || v.trim().isEmpty) return 'Email is required';
-                        if (!RegExp(r'^[^@\s]+@[^@\s]+\.[^@\s]+$').hasMatch(v.trim())) return 'Enter a valid email';
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 14),
-                    _Field(label: 'Password', controller: _passwordController, obscure: true,
-                      validator: (v){
-                        if (v == null || v.isEmpty) return 'Password is required';
-                        if (v.length < 6) return 'Min 6 characters';
-                        return null;
-                      },
-                    ),
-                    const SizedBox(height: 14),
-                    _Field(label: 'Re–enter Password', controller: _confirmPasswordController, obscure: true,
-                      validator: (v){
-                        if (v != _passwordController.text) return 'Passwords do not match';
-                        return null;
-                      },
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 20),
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  onPressed: _submitting ? null : _registerWithEmail,
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: const Color(0xFF0E2B1C),
-                    foregroundColor: const Color(0xFFEDFDDE),
-                    padding: const EdgeInsets.symmetric(vertical: 16),
-                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-                  ),
-                  child: _submitting ? const SizedBox(height: 20, width: 20, child: CircularProgressIndicator(strokeWidth: 2, color: Colors.white)) : const Text('continue', style: TextStyle(fontWeight: FontWeight.w800, fontSize: 18)),
-                ),
-              ),
-              const SizedBox(height: 16),
-              Center(child: Text('Or signup with', style: TextStyle(color: Colors.white.withOpacity(0.7)))),
-              const SizedBox(height: 12),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  _SocialButton(asset: 'assets/social/google.png', onTap: _submitting ? null : _continueWithGoogle),
-                  const SizedBox(width: 18),
-                  _SocialButton(asset: 'assets/social/facebook.png', onTap: null),
-                  const SizedBox(width: 18),
-                  _SocialButton(asset: 'assets/social/instagram.png', onTap: null),
-                ],
-              ),
-              const SizedBox(height: 20),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class EmailVerificationScreen extends StatefulWidget {
-  final String email;
-  const EmailVerificationScreen({super.key, required this.email});
-
-  @override
-  State<EmailVerificationScreen> createState() => _EmailVerificationScreenState();
-}
-
-class _EmailVerificationScreenState extends State<EmailVerificationScreen> {
-  Timer? _pollTimer;
-
-  @override
-  void initState() {
-    super.initState();
-    _startPolling();
-  }
-
-  void _startPolling() {
-    _pollTimer?.cancel();
-    _pollTimer = Timer.periodic(const Duration(seconds: 3), (_) async {
-      final user = FirebaseAuth.instance.currentUser;
-      await user?.reload();
-      if (user != null && user.emailVerified) {
-        _pollTimer?.cancel();
-        if (!mounted) return;
-        Navigator.of(context).pushAndRemoveUntil(
-          _slideRoute(const UsernameScreen()),
-          (route) => false,
-        );
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _pollTimer?.cancel();
-    super.dispose();
-  }
-
-  Future<void> _resend() async {
-    final user = FirebaseAuth.instance.currentUser;
-    await user?.sendEmailVerification();
-    if (!mounted) return;
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('Verification email sent')),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      appBar: AppBar(
-        backgroundColor: const Color(0xFF201E1A),
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back, color: Colors.white),
-          onPressed: () => Navigator.of(context).pop(),
-        ),
-      ),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-        child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 8),
-              const Text(
-                'Email verification',
-                style: TextStyle(fontSize: 28, fontWeight: FontWeight.w800, color: Colors.white),
-              ),
-              const SizedBox(height: 8),
-              Text('Enter the 4-digit code sent to\n${widget.email}',
-                  style: TextStyle(color: Colors.white.withOpacity(0.7))),
-              const SizedBox(height: 24),
-              Row(
-                children: List.generate(4, (index) {
-                  return Container(
-                    margin: EdgeInsets.only(right: index == 3 ? 0 : 12),
-                    height: 56,
-                    width: 56,
-                    decoration: BoxDecoration(
-                      color: const Color(0xFF201E1A),
-                      borderRadius: BorderRadius.circular(16),
-                      border: Border.all(color: Colors.white.withOpacity(0.4)),
-                    ),
-                  );
-                }),
-              ),
-              const SizedBox(height: 12),
-              Row(
-          children: <Widget>[
-                  Text("Didn't get the code? ", style: TextStyle(color: Colors.white.withOpacity(0.6), fontSize: 12)),
-                  GestureDetector(
-                    onTap: _resend,
-                    child: const Text('Resend Code', style: TextStyle(color: Colors.white, fontWeight: FontWeight.w600, fontSize: 12)),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _Field extends StatelessWidget {
-  final String label;
-  final TextEditingController controller;
-  final TextInputType? keyboardType;
-  final bool obscure;
-  final String? Function(String?)? validator;
-
-  const _Field({required this.label, required this.controller, this.keyboardType, this.obscure = false, this.validator});
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: controller,
-      keyboardType: keyboardType,
-      obscureText: obscure,
-      style: const TextStyle(color: Colors.white),
-      validator: validator ?? (v) => (v == null || v.trim().isEmpty) ? '$label is required' : null,
-      decoration: InputDecoration(
-        labelText: label,
-        labelStyle: TextStyle(color: Colors.white.withOpacity(0.7), fontSize: 12),
-        filled: true,
-        fillColor: const Color(0xFF201E1A),
-        enabledBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(16),
-          borderSide: BorderSide(color: Colors.white.withOpacity(0.25)),
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(16),
-          borderSide: const BorderSide(color: Color(0xFFCDEDC6), width: 1.5),
-        ),
-      ),
-    );
-  }
-}
-
-class _SocialButton extends StatelessWidget {
-  final String asset;
-  final VoidCallback? onTap;
-  const _SocialButton({required this.asset, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    final Widget icon = Image.asset(asset, height: 36, width: 36);
-    return GestureDetector(onTap: onTap, child: icon);
-  }
-}
-
-class _PreferenceItem {
-  final String label;
-  final String assetPath;
-  const _PreferenceItem({required this.label, required this.assetPath});
-}
-
-class _PreferenceTile extends StatelessWidget {
-  final _PreferenceItem item;
-  final bool selected;
-  final VoidCallback onTap;
-
-  const _PreferenceTile({required this.item, required this.selected, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Column(
-        children: <Widget>[
-          Container(
-            height: 78,
-            width: 78,
-            decoration: BoxDecoration(
-              shape: BoxShape.circle,
-              color: selected ? const Color(0xFFEDFDDE) : const Color(0xFF2A2723),
-              border: Border.all(
-                color: selected ? const Color(0xFFCDEDC6) : Colors.white.withOpacity(0.15),
-                width: 1.5,
-              ),
-            ),
-            alignment: Alignment.center,
-            child: Padding(
-              padding: const EdgeInsets.all(12.0),
-              child: Image.asset(
-                item.assetPath,
-                fit: BoxFit.contain,
-              ),
-            ),
-          ),
-          const SizedBox(height: 6),
-          Text(
-            item.label,
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              color: Colors.white.withOpacity(0.9),
-              fontSize: 11,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class UsernameScreen extends StatefulWidget {
-  const UsernameScreen({super.key});
-
-  @override
-  State<UsernameScreen> createState() => _UsernameScreenState();
-}
-
-class _UsernameScreenState extends State<UsernameScreen> {
-  final TextEditingController _usernameController = TextEditingController();
-  final _formKey = GlobalKey<FormState>();
-  bool _submitting = false;
-
-  Future<void> _saveUsername() async {
-    if (!_formKey.currentState!.validate()) return;
-    
-    setState(() => _submitting = true);
-    try {
-      final username = _usernameController.text.trim().toLowerCase();
-      final user = FirebaseAuth.instance.currentUser;
-      
-      if (user == null) {
-        throw Exception('User not authenticated');
-      }
-
-      // Check if username already exists
-      final usernameQuery = await FirebaseFirestore.instance
-          .collection('usernames')
-          .doc(username)
-          .get();
-      
-      if (usernameQuery.exists) {
-        setState(() => _submitting = false);
-        if (!mounted) return;
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(content: Text('Username already taken')),
-        );
-        return;
-      }
-
-      // Save username to usernames collection
-      await FirebaseFirestore.instance
-          .collection('usernames')
-          .doc(username)
-          .set({'userId': user.uid, 'createdAt': FieldValue.serverTimestamp()});
-
-      // Save user profile
-      await FirebaseFirestore.instance
-          .collection('users')
-          .doc(user.uid)
-          .set({
-        'username': username,
-        'email': user.email,
-        'displayName': user.displayName ?? '',
-        'createdAt': FieldValue.serverTimestamp(),
-      }, SetOptions(merge: true));
-
-      if (!mounted) return;
-      Navigator.of(context).pushReplacement(
-        _slideRoute(const ContentPreferenceScreen()),
-      );
-    } catch (e) {
-      setState(() => _submitting = false);
-      if (!mounted) return;
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Error: ${e.toString()}')),
-      );
-    }
-  }
-
-  @override
-  void dispose() {
-    _usernameController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF201E1A),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              const SizedBox(height: 16),
-              const Text(
-                'Choose a username',
-                style: TextStyle(fontSize: 32, fontWeight: FontWeight.w800, color: Colors.white),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                "This will be your unique identity in HealthLab. You can use letters, numbers, or underscores.",
-                style: TextStyle(color: Colors.white.withOpacity(0.7), fontSize: 14, height: 1.5),
-              ),
-              const SizedBox(height: 32),
-              Form(
-                key: _formKey,
-                child: TextFormField(
-                  controller: _usernameController,
-                  style: const TextStyle(color: Colors.white),
-                  validator: (value) {
-                    if (value == null || value.trim().isEmpty) {
-                      return 'Username is required';
-                    }
-                    final username = value.trim();
-                    if (username.length < 3) {
-                      return 'Username must be at least 3 characters';
-                    }
-                    if (username.length > 20) {
-                      return 'Username must be less than 20 characters';
-                    }
-                    if (!RegExp(r'^[a-zA-Z0-9_]+$').hasMatch(username)) {
-                      return 'Only letters, numbers, and underscores allowed';
-                    }
-                    return null;
-                  },
-                  decoration: InputDecoration(
-                    hintText: 'Enter username',
-                    hintStyle: TextStyle(color: Colors.white.withOpacity(0.5)),
-                    filled: true,
-                    fillColor: const Color(0xFF201E1A),
-                    enabledBorder: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(16),
-                      borderSide: BorderSide(color: Colors.white.withOpacity(0.25)),
-                    ),
-                    focusedBorder: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(16),
-                      borderSide: const BorderSide(color: Color(0xFFCDEDC6), width: 1.5),
-                    ),
-                    errorBorder: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(16),
-                      borderSide: const BorderSide(color: Colors.red, width: 1.5),
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 32),
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  onPressed: _submitting ? null : _saveUsername,
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: const Color(0xFF0E2B1C),
-                    foregroundColor: const Color(0xFFEDFDDE),
-                    padding: const EdgeInsets.symmetric(vertical: 16),
-                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-                  ),
-                  child: _submitting 
-                    ? const SizedBox(
-                        height: 20, 
-                        width: 20, 
-                        child: CircularProgressIndicator(strokeWidth: 2, color: Colors.white)
-                      ) 
-                    : const Text('continue', style: TextStyle(fontWeight: FontWeight.w800, fontSize: 18)),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
