+++ lib/screens/sign_up_screen.dart
@@ -1,0 +1,211 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_sign_in/google_sign_in.dart';
+import '../utils/app_utils.dart';
+import 'email_verification_screen.dart';
+import 'username_screen.dart';
+
+class SignUpScreen extends StatefulWidget {
+  const SignUpScreen({super.key});
+
+  @override
+  State<SignUpScreen> createState() => _SignUpScreenState();
+}
+
+class _SignUpScreenState extends State<SignUpScreen> {
+  final _formKey = GlobalKey<FormState>();
+  final TextEditingController _nameController = TextEditingController();
+  final TextEditingController _emailController = TextEditingController();
+  final TextEditingController _passwordController = TextEditingController();
+  final TextEditingController _confirmPasswordController = TextEditingController();
+
+  bool _submitting = false;
+
+  Future<void> _registerWithEmail() async {
+    if (!_formKey.currentState!.validate()) return;
+    setState(() => _submitting = true);
+    try {
+      final email = _emailController.text.trim();
+      final password = _passwordController.text;
+      await FirebaseAuth.instance.createUserWithEmailAndPassword(email: email, password: password);
+      final user = FirebaseAuth.instance.currentUser;
+      await user?.sendEmailVerification();
+      if (!mounted) return;
+      Navigator.of(context).pushReplacement(
+        slideRoute(EmailVerificationScreen(email: email)),
+      );
+    } on FirebaseAuthException catch (e) {
+      if (!mounted) return;
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Sign up failed')));
+    } finally {
+      if (mounted) setState(() => _submitting = false);
+    }
+  }
+
+  Future<void> _continueWithGoogle() async {
+    setState(() => _submitting = true);
+    try {
+      final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
+      if (googleUser == null) {
+        setState(() => _submitting = false);
+        return; // canceled
+      }
+      final GoogleSignInAuthentication googleAuth = await googleUser.authentication;
+      final credential = GoogleAuthProvider.credential(
+        accessToken: googleAuth.accessToken,
+        idToken: googleAuth.idToken,
+      );
+      await FirebaseAuth.instance.signInWithCredential(credential);
+      if (!mounted) return;
+      Navigator.of(context).pushReplacement(slideRoute(const UsernameScreen()));
+    } on FirebaseAuthException catch (e) {
+      if (!mounted) return;
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Google sign-in failed')));
+    } finally {
+      if (mounted) setState(() => _submitting = false);
+    }
+  }
+
+  @override
+  void dispose() {
+    _nameController.dispose();
+    _emailController.dispose();
+    _passwordController.dispose();
+    _confirmPasswordController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: AppColors.primaryBackground,
+      body: SafeArea(
+        child: SingleChildScrollView(
+          padding: const EdgeInsets.symmetric(horizontal: 24.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              const SizedBox(height: 16),
+              const Text(
+                'Welcome!',
+                style: TextStyle(fontSize: 32, fontWeight: FontWeight.w800, color: Colors.white),
+              ),
+              const SizedBox(height: 4),
+              Text(
+                "Let's create you an account",
+                style: TextStyle(color: Colors.white.withOpacity(0.7)),
+              ),
+              const SizedBox(height: 24),
+              Form(
+                key: _formKey,
+                child: Column(
+                  children: <Widget>[
+                    _Field(label: 'Full name', controller: _nameController, keyboardType: TextInputType.name),
+                    const SizedBox(height: 14),
+                    _Field(label: 'Email', controller: _emailController, keyboardType: TextInputType.emailAddress,
+                      validator: (v){
+                        if (v == null || v.trim().isEmpty) return 'Email is required';
+                        if (!RegExp(r'^[^@\s]+@[^@\s]+\.[^@\s]+$').hasMatch(v.trim())) return 'Enter a valid email';
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 14),
+                    _Field(label: 'Password', controller: _passwordController, obscure: true,
+                      validator: (v){
+                        if (v == null || v.isEmpty) return 'Password is required';
+                        if (v.length < 6) return 'Min 6 characters';
+                        return null;
+                      },
+                    ),
+                    const SizedBox(height: 14),
+                    _Field(label: 'Re–enter Password', controller: _confirmPasswordController, obscure: true,
+                      validator: (v){
+                        if (v != _passwordController.text) return 'Passwords do not match';
+                        return null;
+                      },
+                    ),
+                  ],
+                ),
+              ),
+              const SizedBox(height: 20),
+              SizedBox(
+                width: double.infinity,
+                child: ElevatedButton(
+                  onPressed: _submitting ? null : _registerWithEmail,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: AppColors.darkGreen,
+                    foregroundColor: AppColors.lightGreen,
+                    padding: const EdgeInsets.symmetric(vertical: 16),
+                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+                  ),
+                  child: _submitting ? const SizedBox(height: 20, width: 20, child: CircularProgressIndicator(strokeWidth: 2, color: Colors.white)) : const Text('continue', style: TextStyle(fontWeight: FontWeight.w800, fontSize: 18)),
+                ),
+              ),
+              const SizedBox(height: 16),
+              Center(child: Text('Or signup with', style: TextStyle(color: Colors.white.withOpacity(0.7)))),
+              const SizedBox(height: 12),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  _SocialButton(asset: 'assets/social/google.png', onTap: _submitting ? null : _continueWithGoogle),
+                  const SizedBox(width: 18),
+                  _SocialButton(asset: 'assets/social/facebook.png', onTap: null),
+                  const SizedBox(width: 18),
+                  _SocialButton(asset: 'assets/social/instagram.png', onTap: null),
+                ],
+              ),
+              const SizedBox(height: 20),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _Field extends StatelessWidget {
+  final String label;
+  final TextEditingController controller;
+  final TextInputType? keyboardType;
+  final bool obscure;
+  final String? Function(String?)? validator;
+
+  const _Field({required this.label, required this.controller, this.keyboardType, this.obscure = false, this.validator});
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      controller: controller,
+      keyboardType: keyboardType,
+      obscureText: obscure,
+      style: const TextStyle(color: Colors.white),
+      validator: validator ?? (v) => (v == null || v.trim().isEmpty) ? '$label is required' : null,
+      decoration: InputDecoration(
+        labelText: label,
+        labelStyle: TextStyle(color: Colors.white.withOpacity(0.7), fontSize: 12),
+        filled: true,
+        fillColor: AppColors.primaryBackground,
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(16),
+          borderSide: BorderSide(color: Colors.white.withOpacity(0.25)),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(16),
+          borderSide: const BorderSide(color: AppColors.primaryGreen, width: 1.5),
+        ),
+      ),
+    );
+  }
+}
+
+class _SocialButton extends StatelessWidget {
+  final String asset;
+  final VoidCallback? onTap;
+  const _SocialButton({required this.asset, required this.onTap});
+
+  @override
+  Widget build(BuildContext context) {
+    final Widget icon = Image.asset(asset, height: 36, width: 36);
+    return GestureDetector(onTap: onTap, child: icon);
+  }
+}
