+++ lib/screens/splash_screen.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'onboarding_screen.dart';
+
+class SplashScreen extends StatefulWidget {
+  const SplashScreen({super.key});
+
+  @override
+  State<SplashScreen> createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  void initState() {
+    super.initState();
+    Future.delayed(const Duration(milliseconds: 1500), () {
+      if (!mounted) return;
+      Navigator.of(context).pushReplacement(
+        MaterialPageRoute(builder: (_) => const OnboardingScreen()),
+      );
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      backgroundColor: Color(0xFF201E1A),
+      body: Center(
+        child: Image(
+          image: AssetImage('assets/logo.png'),
+          fit: BoxFit.contain,
+        ),
+      ),
+    );
+  }
+}
